+++ packages/terminal/src/browser/terminal-frontend-contribution.ts
@@ -179,7 +179,7 @@ export class TerminalFrontendContribution implements FrontendApplicationContribu
 
     @postConstruct()
     protected init(): void {
-        this.shell.currentChanged.connect(() => this.updateCurrentTerminal());
+        this.shell.onDidChangeCurrentWidget(() => this.updateCurrentTerminal());
         this.widgetManager.onDidCreateWidget(({ widget }) => {
             if (widget instanceof TerminalWidget) {
                 this.updateCurrentTerminal();
@@ -191,7 +191,7 @@ export class TerminalFrontendContribution implements FrontendApplicationContribu
         const terminalFocusKey = this.contextKeyService.createKey<boolean>('terminalFocus', false);
         const updateFocusKey = () => terminalFocusKey.set(this.shell.activeWidget instanceof TerminalWidget);
         updateFocusKey();
-        this.shell.activeChanged.connect(updateFocusKey);
+        this.shell.onDidChangeActiveWidget(updateFocusKey);
 
         this.terminalWatcher.onStoreTerminalEnvVariablesRequested(data => {
             this.storageService.setData(ENVIRONMENT_VARIABLE_COLLECTIONS_KEY, data);